requirements.txt
flask
langchain-community
langchain-chroma
pypdf
langchainhub
transformers
langchain-google-genai



-e .